+++ crates/full-node/sov-stf-runner/tests/integration/helpers/hash_stf.rs
@@ -9,7 +9,7 @@ use sov_rollup_interface::common::RollupHeight;
 use sov_rollup_interface::da::{BlobReaderTrait, BlockHeaderTrait, DaSpec, RelevantBlobIters};
 use sov_rollup_interface::stf::{ApplySlotOutput, GenesisParams, StateTransitionFunction};
 use sov_rollup_interface::zk::aggregated_proof::SerializedAggregatedProof;
-use sov_rollup_interface::zk::{ZkVerifier, Zkvm};
+use sov_rollup_interface::zk::ZkVerifier;
 use sov_state::namespaces::User;
 use sov_state::nomt::prover_storage::NomtProverStorage;
 use sov_state::storage::{NativeStorage, SlotKey, SlotValue};
@@ -91,9 +91,7 @@ impl From<Vec<u8>> for HashStfGenesisParams {
     }
 }
 
-impl<InnerVm: Zkvm, OuterVm: Zkvm, Da: DaSpec> StateTransitionFunction<InnerVm, OuterVm, Da>
-    for HashStf
-{
+impl<Da: DaSpec> StateTransitionFunction<Da> for HashStf {
     type Address = MockAddress;
     type StateRoot = StorageRoot<S>;
     type GenesisParams = HashStfGenesisParams;
@@ -132,7 +130,7 @@ impl<InnerVm: Zkvm, OuterVm: Zkvm, Da: DaSpec> StateTransitionFunction<InnerVm, 
         slot_header: &Da::BlockHeader,
         relevant_blobs: RelevantBlobIters<&mut [Da::BlobTransaction]>,
         _execution_context: sov_modules_api::ExecutionContext,
-    ) -> ApplySlotOutput<InnerVm, OuterVm, Da, Self> {
+    ) -> ApplySlotOutput<Da, Self> {
         // Note: Uses native code, so won't work in ZK
         let storage_root_hash = pre_state
             .get_latest_root_hash()
@@ -212,7 +210,7 @@ impl<InnerVm: Zkvm, OuterVm: Zkvm, Da: DaSpec> StateTransitionFunction<InnerVm, 
             "Post apply slot root hashes",
         );
 
-        ApplySlotOutput::<InnerVm, OuterVm, Da, Self> {
+        ApplySlotOutput::<Da, Self> {
             state_root,
             change_set,
             proof_receipts,