+++ examples/demo-rollup/provers/sp1/guest-celestia/src/main.rs
@@ -32,8 +32,7 @@ pub fn main() {
         rollup_proof_namespace: ROLLUP_PROOF_NAMESPACE,
     };
 
-    let stf_verifier =
-        StfVerifier::<_, _, _, SP1, MockZkvm>::new(stf, CelestiaVerifier::new(rollup_params));
+    let stf_verifier = StfVerifier::new(stf, CelestiaVerifier::new(rollup_params));
 
     stf_verifier
         .run_block(guest, storage)